pyproject.toml
[project]
name = "near-omni-client"
version = "0.0.1"
description = "Minimal Python library to interact with the NEAR blockchain and Chain Signatures"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "base58>=2.1.1",
    "ecdsa>=0.19.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "py-near-primitives>=0.2.3",
    "pydantic>=2.11.3",
    "pynacl>=1.5.0",
    "secp256k1>=0.14.0",
]

[project.optional-dependencies]
docs = [
    "furo>=2024.8.6",
    "myst-parser>=4.0.1",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-autodoc-typehints>=3.1.0",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxawesome-theme>=5.3.2",
    "sphinxcontrib-napoleon>=0.7",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
 "pytest-asyncio>=0.26.0",
 "python-dotenv>=1.1.0",
 "ruff>=0.11.6",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["./build/**", "./dist/**", "./docs/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "D", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["D213", "D203", "D100", "D104", "D107", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
